# gxswagger/api/src/main/java/com/jztey/exam/domain/User.java
package com.jztey.exam.domain;

import java.io.Serializable;

public class User implements Serializable {
	private long id;
	private String name;
	private String sex;

	public long getId() {
		return id;
	}

	public void setId(long id) {
		this.id = id;
	}

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}

	public String getSex() {
		return sex;
	}

	public void setSex(String sex) {
		this.sex = sex;
	}

}